A few portability changes, now builds on OSX without warnings.

+++ src/iperf_util.c
@@ -48,7 +48,7 @@ make_cookie(char *cookie)
     /* Generate a string based on hostname, time, randomness, and filler. */
     (void) gethostname(hostname, sizeof(hostname));
     (void) gettimeofday(&tv, 0);
-    (void) snprintf(temp, sizeof(temp), "%s.%ld.%06ld.%08lx%08lx.%s", hostname, tv.tv_sec, tv.tv_usec, (unsigned long int) random(), (unsigned long int) random(), "1234567890123456789012345678901234567890");
+    (void) snprintf(temp, sizeof(temp), "%s.%ld.%06ld.%08lx%08lx.%s", hostname, (unsigned long int) tv.tv_sec, (unsigned long int) tv.tv_usec, (unsigned long int) random(), (unsigned long int) random(), "1234567890123456789012345678901234567890");
 
     /* Now truncate it to 36 bytes and terminate. */
     memcpy(cookie, temp, 36);
